+++ mosaicizers/mosaic-v2.py
@@ -1,247 +1,30 @@
-import sys
+###############################################
+# This file will create a mosaic from CLI.
+# TODO need to sort out functions here
+# In a new extension of Mosaic.
+###############################################
+
 import os
-import io
-import hashlib
 import os.path
 import argparse
-from PIL import Image, ImageOps
-from scipy.spatial import KDTree
-import numpy as np
-from skimage.metrics import structural_similarity as ssim
 from multiprocessing import Process, Queue, cpu_count
 
-# Change these 3 config parameters to suit your needs...
+# These are the custom imports
+import MosaicImage
+import TargetImage
+import TileProcessor
+import TileFitterSciKit
+
+# These are now configed by CLI or class defaults
 TILE_SIZE = 50     # height/width of mosaic tiles in pixels
 TILE_MATCH_RES = 5      # tile matching resolution
-ENLARGEMENT = 8      # mosaic image will be this many times larger
-Image.MAX_IMAGE_PIXELS = None  # Dangerous, but allow it for now
+# ENLARGEMENT = 8      # mosaic image will be this many times larger
+# Image.MAX_IMAGE_PIXELS = None  # Dangerous, but allow it for now
 
 TILE_BLOCK_SIZE = TILE_SIZE / max(min(TILE_MATCH_RES, TILE_SIZE), 1)
 WORKER_COUNT = max(cpu_count() - 1, 1)
-OUT_FILE = 'mosaic.jpeg'
+# OUT_FILE = 'mosaic.jpeg'
 EOQ_VALUE = None
-
-
-class TileProcessor:
-    def __init__(self, tiles_directory):
-        self.tiles_directory = tiles_directory
-
-    def get_average_color(self, img_path_or_obj):
-        """
-        The 'Resize Trick': Shrinks an image to 1x1 to find the mean RGB.
-        Accepts either a file path or an existing PIL Image object.
-        """
-        try:
-            # If it's a path, open it; otherwise assume it's an Image object
-            if isinstance(img_path_or_obj, str):
-                img = Image.open(img_path_or_obj)
-            else:
-                img = img_path_or_obj
-
-            img = img.convert('RGB')
-            # BOX resampling is fast and mathematically accurate for averaging
-            img_tiny = img.resize((1, 1), resample=Image.Resampling.BOX)
-            return img_tiny.getpixel((0, 0))
-        except Exception as e:
-            print(f"Error processing image: {e}")
-            return None
-
-    def __process_tile(self, tile_path):
-        try:
-            img = Image.open(tile_path)
-            img = ImageOps.exif_transpose(img)
-
-            # tiles must be square, so get the largest square that fits inside
-            w = img.size[0]
-            h = img.size[1]
-            min_dimension = min(w, h)
-            w_crop = (w - min_dimension) / 2
-            h_crop = (h - min_dimension) / 2
-            img = img.crop((w_crop, h_crop, w - w_crop, h - h_crop))
-
-            large_tile_img = img.resize((TILE_SIZE, TILE_SIZE), Image.LANCZOS)
-            small_tile_img = img.resize(
-                (int(TILE_SIZE / TILE_BLOCK_SIZE),
-                 int(TILE_SIZE / TILE_BLOCK_SIZE)),
-                Image.LANCZOS
-            )
-
-            return (large_tile_img.convert('RGB'),
-                    small_tile_img.convert('RGB'))
-        except Exception:
-            return (None, None)
-
-    def get_tiles(self):
-        large_tiles = []
-        small_tiles = []
-        count = 0
-        exp_threshold = 1  # for logging
-        print('Reading tiles from {}...'.format(self.tiles_directory))
-
-        # search the tiles directory recursively
-        for root, subFolders, files in os.walk(self.tiles_directory):
-            for tile_name in files:
-                tile_path = os.path.join(root, tile_name)
-                large_tile, small_tile = self.__process_tile(tile_path)
-                if large_tile:
-                    large_tiles.append(large_tile)
-                    small_tiles.append(small_tile)
-                count += 1
-                if count == exp_threshold:
-                    print(f'Processed {count} file(s) so far...')
-                    exp_threshold = exp_threshold * 2
-
-        print('Processed {} tiles.'.format(len(large_tiles)))
-
-        return (large_tiles, small_tiles)
-
-
-class TargetImage:
-    def __init__(self, image_path):
-        self.image_path = image_path
-
-    def get_data(self):
-        print('Processing main image...')
-        img = Image.open(self.image_path)
-        w = img.size[0] * ENLARGEMENT
-        h = img.size[1] * ENLARGEMENT
-        large_img = img.resize((w, h), Image.LANCZOS)
-        w_diff = (w % TILE_SIZE) / 2
-        h_diff = (h % TILE_SIZE) / 2
-
-        # crop the image slightly so we use a whole number of tiles
-        if w_diff or h_diff:
-            large_img = large_img.crop(
-                      (w_diff, h_diff, w - w_diff, h - h_diff)
-            )
-
-        small_img = large_img.resize(
-            (int(w / TILE_BLOCK_SIZE), int(h / TILE_BLOCK_SIZE)),
-            Image.LANCZOS
-        )
-
-        image_data = (large_img.convert('RGB'), small_img.convert('RGB'))
-
-        print('Main image processed.')
-
-        return image_data
-
-
-class TileFitter:
-    def __init__(self, tiles_data):
-        self.tiles_data = tiles_data
-
-    def get_average_color(self, img_path_or_obj):
-        try:
-            # If it's a path, open it; otherwise assume it's an Image object
-            if isinstance(img_path_or_obj, str):
-                img = Image.open(img_path_or_obj)
-            else:
-                img = img_path_or_obj
-
-            img = img.convert('RGB')
-            # BOX resampling is fast and mathematically accurate for averaging
-            img_tiny = img.resize((1, 1), resample=Image.Resampling.BOX)
-            return img_tiny.getpixel((0, 0))
-        except Exception as e:
-            print(f"Error processing image: {e}")
-            return None
-
-    def __get_tile_diff(self, t1, t2, bail_out_value):
-        diff = 0
-        for i in range(len(t1)):
-            diff += ((t1[i][0] - t2[i][0])**2 +
-                     (t1[i][1] - t2[i][1])**2 +
-                     (t1[i][2] - t2[i][2])**2)
-            if diff > bail_out_value:
-                return diff
-        return diff
-
-    def get_best_fit_tile(self, img_data):
-        """ Image Data is the from the Sample """
-        best_fit_tile_index = None
-        min_diff = sys.maxsize
-        tile_index = 0
-
-        for tile_data in self.tiles_data:
-            diff = self.__get_tile_diff(img_data, tile_data, min_diff)
-            if diff < min_diff:
-                min_diff = diff
-                best_fit_tile_index = tile_index
-            tile_index += 1
-
-        return best_fit_tile_index
-
-
-class TileFitterSciKit:
-    def __init__(self, tiles_data, match_res=TILE_MATCH_RES, penalty=0.02):
-        # tiles_data here is the 'small_tiles' list from TileProcessor
-        self.penalty = penalty
-        self.usages = [0.0 for x in range(len(tiles_data))]
-        self.tiles_data = tiles_data
-        self.match_res = match_res
-
-        print("Initializing KDTree for hybrid search...")
-        # 1. Convert tiles to NumPy arrays once
-        # We reshape them from flat lists back into (5x5x3) blocks for SSIM
-        self.tiles_np = [
-            np.array(t).reshape((self.match_res, self.match_res, 3))
-            for t in tiles_data
-        ]
-
-        # 2. Pre-calculate average colors for the Tree
-        avg_colors = [t.mean(axis=(0, 1)) for t in self.tiles_np]
-        self.tree = KDTree(np.array(avg_colors))
-        print("KDTree + SSIM Hybrid Fitter Ready.")
-
-    def get_best_fit_tile(self, img_data):
-        """
-        img_data: A flat list of pixels (from original code's getdata())
-        We convert it to NumPy to use the Tree and SSIM.
-        """
-        # Convert the incoming list to a 5x5x3 array
-        target_np = np.array(img_data).reshape(
-            (self.match_res, self.match_res, 3))
-
-        # Step 1: KDTree Pruning (The "Bucket" step)
-        # Find the top 40 color matches
-        target_avg = target_np.mean(axis=(0, 1))
-        _, indices = self.tree.query(target_avg, k=40)
-
-        best_score = -1
-        best_fit_tile_index = indices[0]
-
-        # Step 2: SSIM Refinement
-        for idx in indices:
-            candidate_np = self.tiles_np[idx]
-
-            # SSIM needs to know the range of pixel values (0-255)
-            try:
-                score = ssim(target_np,
-                             candidate_np,
-                             channel_axis=2,
-                             data_range=255,
-                             win_size=3)
-                score = score - self.usages[idx]
-
-            except ValueError as e:
-                # codes sometimes breaks and hangs.
-                print(f'Got exception {e} \
-                        skipping index{idx} \
-                        data was {type(candidate_np)}')
-                continue
-
-            if score > best_score:
-                best_score = score
-                best_fit_tile_index = idx
-
-            # Early exit if we find an amazing match
-            if score > 0.98:
-                break
-
-        self.usages[best_fit_tile_index] = \
-            self.usages[best_fit_tile_index] + self.penalty
-        return best_fit_tile_index
 
 
 def fit_tiles(work_queue, result_queue, tiles_data):
@@ -271,47 +54,10 @@ class ProgressCounter:
               flush=True, end='\r')
 
 
-class MosaicImage:
-    def __init__(self, original_img):
-        self.image = Image.new(original_img.mode, original_img.size)
-        self.x_tile_count = int(original_img.size[0] / TILE_SIZE)
-        self.y_tile_count = int(original_img.size[1] / TILE_SIZE)
-        self.total_tiles = self.x_tile_count * self.y_tile_count
-
-    def add_tile(self, tile_data, coords):
-        img = Image.new('RGB', (TILE_SIZE, TILE_SIZE))
-        img.putdata(tile_data)
-        self.image.paste(img, coords)
-
-    def save(self):
-        """
-        Saves the image_obj as a .jpeg to /mnt/ebs/mosaics
-        using its MD5 hash as the filename.
-        """
-        output_dir = "/mnt/ebs/mosaics"
-
-        # Ensure the output directory exists
-        os.makedirs(output_dir, exist_ok=True)
-
-        # 1. Convert image to bytes to calculate hash
-        # We save to a temporary buffer or use the raw data
-        img_byte_arr = io.BytesIO()
-        self.image.save(img_byte_arr, format='JPEG')
-        img_bytes = img_byte_arr.getvalue()
-
-        md5_hash = hashlib.md5(img_bytes).hexdigest()
-
-        filename = f"{md5_hash}.jpeg"
-        final_path = os.path.join(output_dir, filename)
-
-        with open(final_path, "wb") as f:
-            f.write(img_bytes)
-
-        print(f"Mosaic saved to: {final_path}")
-        return final_path
-
-
-def build_mosaic(result_queue, all_tile_data_large, original_img_large):
+def build_mosaic(result_queue,
+                 all_tile_data_large,
+                 original_img_large,
+                 suffix):
     mosaic = MosaicImage(original_img_large)
 
     active_workers = WORKER_COUNT
@@ -330,7 +76,7 @@ def build_mosaic(result_queue, all_tile_data_large, original_img_large):
         except KeyboardInterrupt:
             pass
 
-    OUT_FILEPATH = mosaic.save()
+    OUT_FILEPATH = mosaic.save(suffix=suffix)
     print('\nFinished, output is in', OUT_FILEPATH)
 
 
@@ -399,6 +145,17 @@ def mosaic(img_path, tiles_data):
 
 
 if __name__ == '__main__':
+    def restricted_float(x):
+        try:
+            x = float(x)
+        except ValueError:
+            raise argparse.ArgumentTypeError(f"{x} is not a \
+                                              floating-point number")
+
+        if x < 0.01 or x > 0.5:
+            raise argparse.ArgumentTypeError(f"{x} is not in range \
+                                             [0.01, 0.5]")
+        return x
 
     parser = argparse.ArgumentParser(
         description="Generate a high-quality mosaic.")
@@ -415,6 +172,21 @@ if __name__ == '__main__':
                         help="Path to the directory \
                               containing tiles (default: /mnt/ebs/frames)")
 
+    parser.add_argument("-out_dir", "-o",
+                        default="/mnt/ebs/mosaics",
+                        help="This is the directory the \
+                              Mosaics will be save to.")
+
+    parser.add_argument('-suffix', '-s',
+                        help="Type something here if \
+                              if you want it appended \
+                              to the file name.")
+
+    parser.add_argument('-ratio',
+                        type=restricted_float,
+                        default=0.2,
+                        help="Set the ratio (range: 0.01 to 0.5, \
+                              default: 0.2)")
     args = parser.parse_args()
 
     # Current logic: Only handle the single file mode
